requirements.txt
pydantic_settings
uvicorn
fastapi
ckanapi
python-multipart
jinja2
flake8
pytest
httpx
pytest-mock
trio
pytest-asyncio
psutil
requests>=2.25.0
fastapi-utils
pytest-cov
requests
jupyter>=1.0.0
minio
pymongo>=4.0.0
mongomock>=4.1.0
fastapi-mcp
pelicanfs>=0.1.0
kubernetes
packaging
PyYAML
